Convert to monorepo
This commit converts the repository to a monorepo, using npms
`workspaces`.  We are converting to a monorepo in anticipation of
needing two separate applications for the pdc front-end, the base-field
admin interface and the bulkuploader interface. The PR was modeled off
of the structure of PermanentOrg's stela repo[1]. Applications are
stored in the apps folder, where we have moved our existing vue
code as '@pdc/admin-interface'. Our root level scripts have been
replaced with scripts that can run/lint/test/build all applications
at once.

A key change in project structure is in the introduction of application
level `package.json` files, which we're using to store package
dependencies, while the root level `package.json` can store shared dev
dependencies between applications. All applications are now storing dependencies
in the root level 'node_modules' folder. Similarly, we are now using
application-level typescript config files, which inherit from the root level
typescript file, but importantly direct our tslinting to only the
application files.

We have the option now of using application level ESLint configs, but since
linting will be package-level by default, we only need to introduce
those if we have different linting needs across packages (which we do
not expect to at this time). One change was made to our linting config
due to a documented compatibility issue between the
'import/no-extraneous-dependencies' rule[2] and workspaces. We switched
the devDependecies flag in order to make it comply with dev dependencies
now (potentially) being listed in the root package.json, as opposed to
the package level.

Finally, our github actions and dependabot configs have been updated to
reference specific application using the `-w` flag. Additionally, updates
in dependabot have been moved back to a weekly cadence, as part of this
repository moving back to a more active development schedule.

[1] https://github.com/PermanentOrg/stela [2]
https://github.com/import-js/eslint-plugin-import/issues/1174

Issue #1048 Convert to a monorepo