Create Bulk-Uploader application
This commit creates the bulk-uploader application, which is beginning as
an exact copy of the admin-interface application. Though the final form
of the application will be much different than the existing code, it is
likely that most of the code will remain in the form of a login screen
(which is really all that the vue code does at the moment). We did well
to configure the monorepo to support multiple applications, so the only
specific reference we need to update is in dependabot. Otherwise the
root scripts and workspaces in the root package.json cover the new
application. Also, a very minor change included in this commit is the
correcting the errounesoly capitalized 'Monorepo' in the README.md.

Issue #1069 Create bulk-uploader application