Create components and utilities packages
This commit creates two shared files packges, the 'components' and
'utilties' packages, which are used to abstract the components and
utilities between applications. This is done to avoid redundancy and to
allow for easier development of components moving forward. We'll note
that many of these components will likely not be used in any future
applications, but now we can remove them explicitly from the shared
package as we move forward. Shared styles are now stored in the base.css
file in the shared package, and referenced directly in the imports of
the main.css files of the application. The shared-ui package is neither
built nor published, but instead referenced directly by applications. It
is linted along with applications at the root level.

Issue #1069 Create bulk-uploader application