Create BulkUploader Application
This commit creates the bulkUploader applicaiton as a copy of the admin
interface application. It will take a much different final form, but for now
this serves as an adequate template for the application, especially since
the utilities and shared components have been now abstracted to separate
packages.

Issue #1069 Create BulkUploader Package