Redesign Panel Components
This commit does a moderate redesign to our existing 'panel' components,
which we use as wrappers for our content. They were originally designed
with the intention of being able to render multiple panels on a single
page, using a grid layout, but per our latest design spec it makes more
sense to treat them as the main wrappers for our content. It's still a
useful structure to have for maintaining consistent styles throughout
the apps, but no longer needs many of the accoutrements we'd added
(such as collapsibility and tags). Panels will now have uniform headers that
appear above the content, with room for buttons/links ('header actions').

This refactor also takes the opportunity to solidfy the design of our
landing page, which we are keeping from the previous application. One
small addition to the user flow then is that the side bar now does not
render unless the user has logged in.

Issue #1094 Redesign Panel Components to fit figma