Create table components
This commit adds a suite of table related components, modeled after the
table components that we built in the previous version of this
application for rendering data from the pdc in a row/column format. They
are intended to be used in both the bulk-uploader and base field admin
applications, and are extendable to any data type in the pdc we may wish
to render as such.

Column actions are intended to house future logic for sorting and
filtering the data in tables. Those will be accomplished in a future
commit.

Issue #1068 Implement table components and table filtering functionality