Implement Data Upload Components
This commit adds a suite of data input components designed to be used
across pdc admin applications for handling the upload and management of
data. They use the v-model approach, which is Vue's shorthand for
passing form data between components. There are going to be further
datainput components necessary for the admin interface, but this commit
adds all components that will be necessary for the bulk-uploader in such
a way that they can be mixed and matched across future forms that we
make for user data uploaded to the pdc.

Issue #1093 Create DataInput components