Implement Add Data View
This lengthy commit implements a minimal 'Add Data' view for the
bulk-uploader, providing the first entry point into uploading user data
to the pdc. This involves a suite of helper api hooks that can POST data
to the pdc, following the AWS Presigned Post flow we've adopted. Again,
these hooks were modeled largely off of the work done in the previous
iteration of the application. Similar to the other views, functionality
that is not supported in the Service has been disabled and marked for
future implementation.

Issue #1066 Bulk Uploader: Implement Add Data View