Add Individual Bulk Upload View
This commit adds the view for an individual bulkupload, which shows
the status and further details about the upload. Currently, much
of the desired data for the bulk-upload is not supported in the pdc,
so we have left them as TODOs in the view itself. To that same point,
we are also in need of an individual bulkUploads endpoint to get
one by id. For the time being, we've implemented client-side filtering
across all bulk-uploads, but we can change that once we've implemented
the necessary routes in service.

Issue #1066 Implement Individual Bulk Upload view