Add panel support for rendering data inputs
This commit extends our panel components to support dataInput
components, adding styles and optional props for panels that we specify
as dataInput panels in the form of the variant prop, which currently only supports
data-panel or default, but could feasibly support more variants in the future.
Also, it adds the PanelSection component, which
follows the design pattern specified in our figma for rendering
dataInputs (following a header/input format).  These were opted to be
made part of the panelSection format as we can anticipate future use
cases where we may want to reuse this layout without necessarily
including input fields in the panel sections.

Issue #1093 Create DataInput components