Fix Token initialization in api hook
This commit fixes an issue wherein the user token was not populating at
api call time, leading to unauthorized requests. This was due to the
variable being invoked outside the function scope.  A simple fix was to
invoke the variable from the hook inside of the generic pdcApi function.

Issue #1067 Bulk Uploader: Implement Bulk Uploads page