Remove bulk-upload instructions text for now
This commit removes a placeholder message we had in the bulkuploader,
which was intended to link to more thorough bulk-upload instructions.
Since we are resolving bulk-upload-related informational docs in this
pr, and we do not have those prepared at the time, it would be
preferable to remove the text entirely so as to not cause any confusion.

Issue #1125 Link template CSV in bulk-uploader