Update PresignedPostRequest to File
This commit changes all references from PresignedPostRequests to the new
File entity, as defined in service issue #1689. The functionality of the
two entities are the same, so the only changes needed to be made are in
the naming of our hooks and the route we're calling.

Issue #1127 Update bulk-uploader to use new File entity and route