Update bulk-uploader to support new `file` flow
This commit updates the bulk-uploader component and corresponding api
routes to fit our new `file` upload flow in the service (Issue 1689 in
pdc service). This is mostly the same as our old flow, with the added
feature of passing the file storage key to the bulkUpload task.

Issue #1127 Update bulk-uploader to use new File entity and route