Bump @pdc/sdk to latest version
This commit bumps the @pdc/sdk to the latest version across our apps and
packages. Considering that all our apps/packages do employ the pdc/sdk,
and further that code in this repo is fundemtnally intended to interface
with the pdc service, it may make sense to make this a top level
package, but that could be left to a future commit.

Issue #1127 Update bulk-uploader to use new File entity and route