Update BulkUpload tasks to reference file entities
This commit updates our bulkuploader to fit the latest service spec,
such that bulkUploadTasks now reference file entities directly in their
creation. It also updates the sdk to the latest version across our apps
and packages.

It also removes a (currently unused) tag in `/BulkUploadView`, which was
intended as a placeholder for downloading the original file of a
bulkupload. It accessed a now-deprecated field on the bulkUploadTask
entity, but importantly also highlighted an error in our swagger-gen:
currently the BulkUploadTask entity has it's 'proposalDataFile'
improperly typed as any. We will need to resolve that on the service
end.

Issue #1140 Update bulkUploadTasks to reference File entities