Add BaseFields view
This commit adds the basefields view, which displays all the basefields
in the pdc instance along with their relevant fields. Currently, they
link to the default basefields page, but will subsequently link to the
individual view for a basefield. The design here is pleasently similar
to our bulkUploads view, and allows for much reuse of components and
structure!

Issue #1159 Add Basefield Admin Basefields view