Create EditBaseFields View
This commit adds in an a view for editing basefield values, using the
new BaseFieldForm component. The idea of the BaseFieldForm component is
to be generalizable to either edit an existing basefield, or create a
new basefield. This is currently set as a boolean flag, though we could
extend this to a behavior ENUM in the future (but I really do not think
we will be having more than two modalities with this form).  As such,
there is a small amount of conditional rendering in the BaseFieldForm.

Furthermore, we're adding two new data input components, one generic
input (TextInput) in the components shared package, and one very
specific rendition on the radio input, the RelevanceDurationInput which
is defined only in the basefield admin.

Since basefield creation is an upsert PUT operation, we can use the same
api hook for creation and updating.

Support for localizations has been decided on being a subsequent PR, as
currently there is no way to create localizations as part of a basefield
PUT. It would have to be chained api calls, and I believe it would be
worth refactoring the service api to accept localization edits as part
of the basefield put before implementing this in the admin.

Issue #1160 Add Basefield Admin Basefield Edit View