Add Add Basefield View
This commit adds a view for creating a new
basefield, which is almost identical to editing a basefield. The only
difference between creating a basefield and editing one, as far as the
ui is concerned, is that when you're creating a basefield, there are no
default values to template into the form inputs, and you are required to
pass a short_code (which, after creation, is not an editable field for
basefields).

Issue #1160 Add Basefield Admin Basefield View