Consolidate dependabot to single package-ecosystem
This commit consolidates our dependabot app/package configurations into
a single config, as per issue #4993 in dependabots repo[1]. We tried the
individual configurations for package updating, but it seems like that
is neither reccomended nor functional, as we were breaking CI/CD with
the latest upgrade of a package. This we believe was caused by the lack
of the 'versioning-strategy increase' value, which is needed in a
monorepo.

[1]https://github.com/dependabot/dependabot-core/issues/4993#issuecomment-1289133027

Issue #1189 Resolve dependabot configuration for monorepos