Switch to KeycloakServerConfig type
As documented in https://github.com/keycloak/keycloak-js/issues/186, the
latest update to keycloak-js refactors the keycloakconfig into a union
type to support both Keycloak and generic openID connect configs, and
this causes a breaking typescript error. Since we're using the keycloak
specific configuration, this commit changes the types accordingly.