Render keycloak username for bulkupload creators
This commit adds an additional fetch on the bulkUploadsView to the users
endpoint, which it uses to get the username from the bulkuploads
`createdBy` field, which is a keycloak user id. It couples the
bulkuploads and users fetches into a single transaction, which seems
like a safer design choice moving forward (we should likely refactor
other fetches in our apps that fetch from multiple endpoints).

Issue #1192 Use keycloakUserName for bulkUpload source