Add File Attachment handling
This commit adds basefield file-attachment handling for the bulkuploader, in the form
of a new file upload panel in the bulkUploader component. It also fixes a duplicate id error
in the file upload component, and splits up the functionality of the addBulkUpload handler (now
that we're adding another layer of complexity).

Issue #1101 Implement File uploader