Enable auto-merge for Dependabot PRs
This new workflow will turn on auto-merge exclusively for PRs that have
been opened by dependabot.  This won't bypass our branch restrictions,
which is critical because it means a person will still need to approve
the PR and CI will still need to pass.

It just saves the reviewer from needing to click the "merge" button in
addition to leaving a review.

Issue #1223 Enable dependabot auto-merge