Add write permission to automerge action
This was part of the sample action we used[1] but I thought it wasn't
necessary after making various modifications.  The resulting workflow is
failing with a permissions error so it seems reasonable to try adding
the permission we removed.

Issue #1237 Dependabot auto-merge fails with "GraphQL: Resource not accessible by integration (enablePullRequestAutoMerge)"

[1] docs.github.com/en/enterprise-cloud@latest/code-security/dependabot/working-with-dependabot/automating-dependabot-with-github-actions#enabling-automerge-on-a-pull-request