Create table sorting composable
This commit adds the lodash sorting package to the repository, and uses
it to create a composable (the first of it's kind in the repo) for
sorting values based on arbitrary fields. It then introduces binding for
the 'useTableSort' functions to tablecolumnHead components, in order to
make these components serve as sorting interfaces for table components
at-large.

Issue #1068 Implement table components and table filtering functionality