Invoke table sorting in table components within our apps
This commit actually adds the sorting features to our table componets in
the bulkuploader and admin interface. Notably, the UI here is a stripped
down version of our proposed final sorting functionality, but this
functions as a stepping stone towards that ultimate design.

Issue #1068 Implement table components and table filtering functionality