Adjust Dependabot cadence
We recently decided to move to a monthly dependabot cadence (as opposed
to weekly) as a way to strike a balance between avoiding dependency
drift and overhead from reviews.

This balance may shift in future, but for now the team has limited time
and we want to be sure we're able to focus on feature development.

This change also adds a 7 day cooldown which means we won't incorporate
releases until after they have existed in the ecosystem for at least one
week

Issue #1291 Modify dependabot cadences