Add tanstack table component and helpers
This commit adds in the tanstack table package, which is a headless
datagrid api that has a wide user base and vue compatibility.  @chriszs
rightly pointed out that as our table needs extend, it would be worth
looking at existing solutions, and specifically for sorting this package
ends up being very friendly for plugging in a wide assortment of data!
This adds a generic 'DataTable' component, which is a departure from our
distributed model of table components, but is more in tune with the tan
stack style. Ideally, we should be able to define one monolithic 'Table'
api and then using options and flags customize it to our needs.

Issue #1068 Implement Table Sorting