Add File Name to bulk upload view
This commit adds the file name to the bulk upload view, and puts a
bandaid over the current sdk issue where the file type is not being
properly referenced. I believe this is because the typescript sdk
rejects types that overload existing types, i.e why the type 'file' gets
build to 'modelFile'. In the meantime, this keeps the type of the actual
field accurate with changes in the service until we fix that bug.

Issue #1240 Add File Metadata